+++ src/lib/useReactDataTable/useReactDataTableProps.tsx
@@ -1,4 +1,4 @@
-﻿import { ColumnDef, ExpandedState, OnChangeFn, PaginationState, RowSelectionState, TableOptions } from "@tanstack/react-table";
+﻿import { ColumnDef, ColumnPinningState, ExpandedState, OnChangeFn, PaginationState, RowSelectionState, TableOptions } from "@tanstack/react-table";
 import { FilterModel, TableState } from "../types/TableState";
 import { SortingState } from "../types/SortingState";
 
@@ -75,6 +75,11 @@ export interface useReactDataTableProps<TData, TFilter extends FilterModel> {
    * event handler for when the expanded changes
    */
   onExpandedChange?: OnChangeFn<ExpandedState>;
+
+  /**
+   * event handler for when the column pinning changes
+   */
+  onColumnPinningChange?:OnChangeFn<ColumnPinningState>;
 
   /**
    * the react table options that will be passed to the `useReactTable` hook.