+++ src/lib/useReactDataTableState/useReactDataTableState.ts
@@ -1,5 +1,5 @@
 ﻿import { useState } from "react";
-import { ExpandedState, PaginationState, RowSelectionState } from "@tanstack/react-table";
+import { ColumnPinningState, ExpandedState, PaginationState, RowSelectionState } from "@tanstack/react-table";
 import { useReactDataTableStateProps } from "./useReactDataTableStateProps";
 import { useReactDataTableStateResult } from "./useReactDataTableStateResult";
 import { FilterModel } from "../types/TableState";
@@ -13,7 +13,7 @@ import { OptionalNullable } from "../types/NullableTypes";
 const useReactDataTableState = <TData, TFilter extends FilterModel = Record<string, never>>(
   props: OptionalNullable<useReactDataTableStateProps<TData, TFilter>>,
 ): useReactDataTableStateResult<TData, TFilter> => {
-  const { initialColumnFilters, initialSorting, initialPagination, initialRowSelection, initialExpanded } =
+  const { initialColumnFilters, initialSorting, initialPagination, initialRowSelection, initialExpanded, initialColumnPinning } =
     props as useReactDataTableStateProps<TData, TFilter>;
 
   const [columnFilters, setColumnFilters] = useState<TFilter>((initialColumnFilters ?? {}) as TFilter);
@@ -21,6 +21,7 @@ const useReactDataTableState = <TData, TFilter extends FilterModel = Record<stri
   const [sorting, setSorting] = useState<SortingState<TData> | undefined>(initialSorting);
   const [rowSelection, setRowSelection] = useState<RowSelectionState>(initialRowSelection ?? ({} as RowSelectionState));
   const [expanded, setExpanded] = useState<ExpandedState>(initialExpanded ?? ({} as ExpandedState));
+  const [columnPinning, setColumnPinning] = useState<ColumnPinningState>(initialColumnPinning ?? ({} as ColumnPinningState));
 
   const [pagination, setPagination] = useState<PaginationState>({
     pageIndex: initialPagination?.pageIndex ?? 0,
@@ -34,12 +35,14 @@ const useReactDataTableState = <TData, TFilter extends FilterModel = Record<stri
     afterSearchFilter,
     rowSelection,
     expanded,
+    columnPinning,
     setSorting,
     setColumnFilters,
     setPagination,
     setAfterSearchFilter,
     setRowSelection,
     setExpanded,
+    setColumnPinning,
   };
 };