+++ src/lib/useReactDataTableState/useReactDataTableStateProps.ts
@@ -1,4 +1,4 @@
-﻿import { ExpandedState, PaginationState, RowSelectionState } from "@tanstack/react-table";
+﻿import { ColumnPinningState, ExpandedState, PaginationState, RowSelectionState } from "@tanstack/react-table";
 import { SortingState } from "../types/SortingState";
 import { FilterModel } from "../types/TableState";
 import { ColumnFilterState } from "../types/ColumnFilterState";
@@ -31,4 +31,9 @@ export interface useReactDataTableStateProps<TData, TFilter extends FilterModel>
    * the initial expanded
    */
   initialExpanded?: ExpandedState;
+
+  /**
+   * the initial column pinning
+   */
+  initialColumnPinning?: ColumnPinningState;
 }